make golangci happy

+++ x/emissions/keeper/actor_utils/util_sort.go
@@ -39,16 +39,16 @@ func FindTopNByScoreDesc(
 	})
 
 	// which is bigger, n or the length of the scores?
-	N := int(n)
-	if N > len(scores) {
-		N = len(scores)
+	N := n
+	if N > uint64(len(scores)) {
+		N = uint64(len(scores))
 	}
 	topNActorsSorted = make([]emissionstypes.Score, N)
 	actorIsTop = make(map[string]struct{}, N)
 	// populate top n actors sorted with only the top n
 	// populate all with all
 	// actor is top is a map of the top n actors
-	for i := 0; i < N; i++ {
+	for i := uint64(0); i < N; i++ {
 		topNActorsSorted[i] = scores[i]
 		actorIsTop[scores[i].Address] = struct{}{}
 	}
